now works with name spaces on real robot as in sim

+++ base_controllers/base_controller_fixed.py
@@ -48,7 +48,6 @@ import matplotlib.pyplot as plt
 
 import  params as conf
 robotName = "ur5"
-real_robot = False
 
 # controller manager management
 from controller_manager_msgs.srv import SwitchControllerRequest, SwitchController
@@ -61,14 +60,17 @@ import actionlib
 
 class BaseController(threading.Thread):
     
-    def __init__(self, robot_name="ur5", real_robot = False):
+    def __init__(self, robot_name="ur5"):
         threading.Thread.__init__(self)
         self.robot_name = robot_name
-        self.real_robot = real_robot
+        self.real_robot = conf.robot_params[self.robot_name]['real_robot']
         self.spawn_x = 0.5
         self.spawn_y = 0.35
         self.spawn_z = 1.8
-        self.use_torque_control = 0
+        if (conf.robot_params[self.robot_name]['control_type'] == "torque"):
+            self.use_torque_control = 1
+        else:
+            self.use_torque_control = 0
 
         # needed to be able to load a custom world file
         print(colored('Adding gazebo model path!', 'blue'))
@@ -77,7 +79,7 @@ class BaseController(threading.Thread):
         if self.real_robot:
             os.system("killall rviz gzserver gzclient")
             print(colored('------------------------------------------------ROBOT IS REAL!', 'blue'))
-            if (not rosgraph.is_master_online()) or ("/ur_hardware_interface"  not in rosnode.get_node_names()):
+            if (not rosgraph.is_master_online()) or ("/" + self.robot_name + "/ur_hardware_interface"  not in rosnode.get_node_names()):
                 print(colored('Error: you need to launch the ur driver!', 'red'))
                 sys.exit()
             else:
@@ -144,24 +146,16 @@ class BaseController(threading.Thread):
         self.verbose = conf.verbose                                                                                           
         self.u.putIntoGlobalParamServer("verbose", self.verbose)
 
-        if (not self.real_robot):
-            self.sub_jstate = ros.Subscriber("/" + self.robot_name + "/joint_states", JointState,
-                                             callback=self._receive_jstate, queue_size=1, tcp_nodelay=True)
+        self.sub_jstate = ros.Subscriber("/" + self.robot_name + "/joint_states", JointState,
+                                         callback=self._receive_jstate, queue_size=1, tcp_nodelay=True)
 
-            self.switch_controller_srv = ros.ServiceProxy(
-                "/" + self.robot_name + "/controller_manager/switch_controller", SwitchController)
-            self.load_controller_srv = ros.ServiceProxy("/" + self.robot_name + "/controller_manager/load_controller",
-                                                        LoadController)
-            self.pub_reduced_des_jstate = ros.Publisher("/" + self.robot_name + "/joint_group_pos_controller/command",
-                                                        Float64MultiArray, queue_size=10)
-        else:
-            # controller manager management (only usable in non torque mode)
-            self.sub_jstate = ros.Subscriber("/joint_states", JointState,
-                                             callback=self._receive_jstate, queue_size=1, tcp_nodelay=True)
+        self.switch_controller_srv = ros.ServiceProxy(
+            "/" + self.robot_name + "/controller_manager/switch_controller", SwitchController)
+        self.load_controller_srv = ros.ServiceProxy("/" + self.robot_name + "/controller_manager/load_controller",
+                                                    LoadController)
+        self.pub_reduced_des_jstate = ros.Publisher("/" + self.robot_name + "/joint_group_pos_controller/command",
+                                                    Float64MultiArray, queue_size=10)
 
-            self.switch_controller_srv = ros.ServiceProxy("/controller_manager/switch_controller", SwitchController)
-            self.load_controller_srv = ros.ServiceProxy("/controller_manager/load_controller", LoadController)
-            self.pub_reduced_des_jstate = ros.Publisher("/joint_group_pos_controller/command", Float64MultiArray, queue_size=10)
 
         if self.real_robot:
             self.available_controllers = [
@@ -203,8 +197,9 @@ class BaseController(threading.Thread):
     def deregister_node(self):
         print( "deregistering nodes"     )
         self.ros_pub.deregister_node()
-        os.system(" rosnode kill /"+self.robot_name+"/ros_impedance_controller")    
-        os.system(" rosnode kill /gazebo")   
+        if not self.real_robot:
+            os.system(" rosnode kill /"+self.robot_name+"/ros_impedance_controller")
+            os.system(" rosnode kill /gzserver /gzclient")
                                                                                                                                      
     def updateKinematicsDynamics(self):
         # q is continuously updated
@@ -291,14 +286,14 @@ class BaseController(threading.Thread):
 
     def send_joint_trajectory(self):
         """Creates a trajectory and sends it using the selected action server"""
-        if not self.real_robot:
+        if self.real_robot:
             trajectory_client = actionlib.SimpleActionClient(
-                "{}/follow_joint_trajectory".format("/ur5/pos_joint_traj_controller"),
+                "{}/follow_joint_trajectory".format("/" + self.robot_name + "/scaled_pos_joint_traj_controller"),
                 FollowJointTrajectoryAction,
             )
         else:
             trajectory_client = actionlib.SimpleActionClient(
-                "{}/follow_joint_trajectory".format("/scaled_pos_joint_traj_controller"),
+                "{}/follow_joint_trajectory".format("/" + self.robot_name + "/pos_joint_traj_controller"),
                 FollowJointTrajectoryAction,
             )
 
@@ -368,49 +363,50 @@ def talker(p):
     p.initVars()     
     p.startupProcedure() 
     ros.sleep(1.0)
-    # to test the trajectory
 
-    if (p.real_robot):
-        p.switch_controller("scaled_pos_joint_traj_controller")
+    #loop frequency
+    rate = ros.Rate(1/conf.robot_params[p.robot_name]['dt'])
+
+    if (conf.robot_params[p.robot_name]['control_mode'] == "trajectory"):
+        # to test the trajectory
+        if (p.real_robot):
+            p.switch_controller("scaled_pos_joint_traj_controller")
+        else:
+            p.switch_controller("pos_joint_traj_controller")
+        p.send_joint_trajectory()
     else:
-        p.switch_controller("pos_joint_traj_controller")
-    p.send_joint_trajectory()
+        #control loop
+        while True:
+            #update the kinematics
+            p.updateKinematicsDynamics()
 
-    #loop frequency       
-    rate = ros.Rate(1/conf.robot_params[p.robot_name]['dt']) 
-                           
-    #control loop
-    while True:  
-        #update the kinematics
-        p.updateKinematicsDynamics()    
-    
-        # set reference            
-        p.q_des  = conf.robot_params[p.robot_name]['q_0']  + 0.1*np.sin(p.time)
+            # set reference
+            p.q_des  = conf.robot_params[p.robot_name]['q_0']  + 0.1*np.sin(p.time)
 
-        # controller 
-        p.tau_ffwd = np.zeros(p.robot.na)       
-        # only torque loop                            
-        #p.tau_ffwd = conf.robot_params[p.robot_name]['kp']*(np.subtract(p.q_des,   p.q))  - conf.robot_params[p.robot_name]['kd']*p.qd     
+            # controller
+            p.tau_ffwd = np.zeros(p.robot.na)
+            # only torque loop
+            #p.tau_ffwd = conf.robot_params[p.robot_name]['kp']*(np.subtract(p.q_des,   p.q))  - conf.robot_params[p.robot_name]['kd']*p.qd
 
-        if (p.use_torque_control):
-            p.send_des_jstate(p.q_des, p.qd_des, p.tau_ffwd)
-            p.ros_pub.add_arrow(p.x_ee, p.contactForceW/(6*p.robot.robot_mass),"green") 
-        else:   
-            p.send_reduced_des_jstate(p.q_des) #no torque fb is present          
+            if (p.use_torque_control):
+                p.send_des_jstate(p.q_des, p.qd_des, p.tau_ffwd)
+                p.ros_pub.add_arrow(p.x_ee, p.contactForceW/(6*p.robot.robot_mass),"green")
+            else:
+                p.send_reduced_des_jstate(p.q_des) #no torque fb is present
 
-        # log variables
-        p.logData()    
-        p.ros_pub.add_marker(p.x_ee)
-        p.ros_pub.publishVisual()                      
-  
-        #wait for synconization of the control loop
-        rate.sleep()     
-       
-        p.time = p.time + conf.robot_params[p.robot_name]['dt']            
-       # stops the while loop if  you prematurely hit CTRL+C                    
-        if ros.is_shutdown():
-            print ("Shutting Down")                    
-            break;                                                
+            # log variables
+            p.logData()
+            p.ros_pub.add_marker(p.x_ee)
+            p.ros_pub.publishVisual()
+
+            #wait for synconization of the control loop
+            rate.sleep()
+
+            p.time = p.time + conf.robot_params[p.robot_name]['dt']
+           # stops the while loop if  you prematurely hit CTRL+C
+            if ros.is_shutdown():
+                print ("Shutting Down")
+                break;
                              
     ros.sleep(1.0)                
     print ("Shutting Down")                 
@@ -420,12 +416,12 @@ def talker(p):
     
 if __name__ == '__main__':
 
-    p = BaseController(robotName, real_robot)
+    p = BaseController(robotName)
     try:
         talker(p)
     except ros.ROSInterruptException:
         from utils.common_functions import plotJoint      
-        plotJoint('position',0, p.time_log, p.q_log, None, p.qd_log, p.qd_des_log, None, None, p.tau_log, p.tau_ffwd_log, p.joint_names)
+        plotJoint('position',0, p.time_log, p.q_log, p.q_des_log, p.qd_log, p.qd_des_log, None, None, p.tau_log, p.tau_ffwd_log, p.joint_names)
         plotJoint('torque',1, p.time_log, p.q_log, p.q_des_log, p.qd_log, p.qd_des_log, None, None, p.tau_log, p.tau_ffwd_log, p.joint_names)
         plt.show(block=True)
